feature/OP-761 - Edit a File BE

+++ app/upload/views.py
@@ -13,8 +13,12 @@ from flask import (
 from werkzeug.utils import secure_filename
 
 from app import upload_redis as redis
-from app.lib.utils import b64decode_lenient
+from app.lib.utils import (
+    b64decode_lenient,
+    eval_request_bool,
+)
 from app.models import Responses
+from app.constants import UPDATED_FILE_DIRNAME
 from app.upload import upload
 from app.upload.constants import (
     CONTENT_RANGE_HEADER,
@@ -38,6 +42,12 @@ def post(request_id):
     For filesize validation and more upload logic, see:
         /static/js/plugins/jquery.fileupload-main.js
 
+    Optional request body parameters:
+    - update (bool)
+        save the uploaded file to the 'updated' directory
+        (this indicates the file is meant to replace
+        a previously uploaded file)
+
     :returns: {
         "name": file name,
         "size": file size
@@ -46,7 +56,8 @@ def post(request_id):
     files = request.files
     file_ = files[next(files.keys())]
     filename = secure_filename(file_.filename)
-    if upload_exists(request_id, filename):
+    is_update = eval_request_bool(request.form.get('update'), False)
+    if not is_update and upload_exists(request_id, filename):
         response = {
             "files": [{
                 "name": filename,
@@ -62,7 +73,7 @@ def post(request_id):
         if not os.path.exists(upload_path):
             os.mkdir(upload_path)
         filepath = os.path.join(upload_path, filename)
-        key = get_upload_key(request_id, filename)
+        key = get_upload_key(request_id, filename, is_update)
 
         try:
             if CONTENT_RANGE_HEADER in request.headers:
@@ -82,13 +93,13 @@ def post(request_id):
                         fp.write(file_.stream.read())
                     # scan if last chunk written
                     if os.path.getsize(filepath) == size:
-                        scan_and_complete_upload.delay(request_id, filepath)
+                        scan_and_complete_upload.delay(request_id, filepath, is_update)
             else:
                 valid_file_type, file_type = is_valid_file_type(file_)
                 if valid_file_type:
                     redis.set(key, upload_status.PROCESSING)
                     file_.save(filepath)
-                    scan_and_complete_upload.delay(request_id, filepath)
+                    scan_and_complete_upload.delay(request_id, filepath, is_update)
 
             if not valid_file_type:
                 response = {
@@ -123,15 +134,18 @@ def post(request_id):
 def delete(r_id_type, r_id, filecode):
     """
     Removes an uploaded file.
-    NOTE: This can only deal with request ids for now (OP-798)
 
     :param r_id_type: "response" or "request"
     :param r_id: the Response or Request identifier
     :param filecode: the encoded name of the uploaded file
         (base64 without padding)
 
-    optional request body parameters:
-    - quarantine_only: only delete the file if it is quarantined
+    Optional request body parameters:
+    - quarantined_only (bool)
+        only delete the file if it is quarantined
+        (beware: takes precedence over 'updated_only')
+    - updated_only (bool)
+        only delete the file if it is in the 'updated' directory
 
     :returns:
         On success:
@@ -141,11 +155,6 @@ def delete(r_id_type, r_id, filecode):
     """
     # TODO: check current user request permissions
     filename = secure_filename(b64decode_lenient(filecode))
-    path_for_status = {
-        upload_status.PROCESSING: current_app.config['UPLOAD_QUARANTINE_DIRECTORY'],
-        upload_status.SCANNING: current_app.config['UPLOAD_QUARANTINE_DIRECTORY'],
-        upload_status.READY: current_app.config['UPLOAD_DIRECTORY']
-    }
     if r_id_type not in ["request", "response"]:
         response = {"error": "Invalid ID type."}
     else:
@@ -153,20 +162,35 @@ def delete(r_id_type, r_id, filecode):
             if r_id_type == "response":
                 response = Responses.query.filter(id=r_id)
                 r_id = response.request_id
-            status = redis.get(
-                get_upload_key(r_id, filename)).decode("utf-8")
-            if status is not None:
-                path = (current_app.config['UPLOAD_QUARANTINE_DIRECTORY']
-                        if request.form.get('quarantine_only', False)
-                        else path_for_status[status])
-                filepath = os.path.join(
-                    os.path.join(path, r_id),
-                    filename)
-                if os.path.exists(filepath):
-                    os.remove(filepath)
-                    response = {"deleted": filename}
-                else:
-                    response = {"error": "File not found."}
+
+            path = ''
+            if eval_request_bool(request.form.get('quarantined_only'), False):
+                path = os.path.join(
+                    current_app.config['UPLOAD_QUARANTINE_DIRECTORY'],
+                    r_id
+                )
+            elif eval_request_bool(request.form.get('updated_only'), False):
+                path = os.path.join(
+                    current_app.config['UPLOAD_DIRECTORY'],
+                    r_id,
+                    UPDATED_FILE_DIRNAME
+                )
+            else:
+                path_for_status = {
+                    upload_status.PROCESSING: current_app.config['UPLOAD_QUARANTINE_DIRECTORY'],
+                    upload_status.SCANNING: current_app.config['UPLOAD_QUARANTINE_DIRECTORY'],
+                    upload_status.READY: current_app.config['UPLOAD_DIRECTORY']
+                }
+                status = redis.get(get_upload_key(r_id, filename))
+                if status is not None:
+                    path = os.path.join(
+                        path_for_status[status.decode("utf-8")],
+                        r_id
+                    )
+            filepath = os.path.join(path, filename)
+            if path != '' and os.path.exists(filepath):
+                os.remove(filepath)
+                response = {"deleted": filename}
             else:
                 response = {"error": "Upload not found."}
         except Exception as e:
@@ -189,6 +213,7 @@ def status():
     Request Parameters:
         - request_id
         - filename
+        - for_update (bool, optional)
 
     :returns: {
         "status": upload status
@@ -198,7 +223,8 @@ def status():
         status = redis.get(
             get_upload_key(
                 request.args['request_id'],
-                secure_filename(request.args['filename'])
+                secure_filename(request.args['filename']),
+                eval_request_bool(request.args.get('for_update'), False)
             )
         )
         if status is not None:
